requirements.txt
flask==3.1.0
functions-framework==3.8.2
google-cloud-vision==3.10.0
python-dotenv==1.0.1